# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https:///www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://kids-collage.test/api",
            "description": "Local Host"
        }
    ],
    "paths": {
...
        "/utility/select/{model}": {
            "get": {
                "tags": [
                    "Utility"
                ],
                "summary": "for use in select box",
                "description": "\n     *     `locker`: available=available= lockers where dosen't have active entry(can assign to new user) | available=unavailable= lockers where have active entry (cannot assign to new user) </br></br>\n     *     `wallet`: type=user= wallets where type is user | type=locker= wallets where type is locker </br></br>",
                "operationId": "selectByModel",
                "parameters": [
                    {
                        "name": "model",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "user",
                            "enum": [
                                "user"
                            ]
                        }
                    },
                    {
                        "name": "options",
                        "in": "query",
                        "description": "you must send as query param param1=1&param2=2 </br>\n     *      `category`:shop_id:int </br></br>\n     *      `wallet`:type=user|locker|contract|organization,charge:1|0,available=available|unavailable,user_id=int",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        },
                        "examples": {
                            "Default": {
                                "summary": "",
                                "value": ""
                            },
                            "search available locker for enter user": {
                                "summary": "",
                                "value": "available=available&locker_group_id=1"
                            },
                            "search unavailable locker for exist user": {
                                "summary": "",
                                "value": "available=unavailable&locker_group_id=1"
                            },
                            "list of categories of a shop": {
                                "summary": "",
                                "value": "shop_id=1"
                            },
                            "lockers in use with charge": {
                                "summary": "",
                                "value": "available=unavailable&charge=1"
                            },
                            "wallet with type:user|locker|contract|organization": {
                                "summary": "",
                                "value": "type=user"
                            },
                            "wallet with type locker can use for product or service payments": {
                                "summary": "",
                                "value": "type=locker&available=unavailable"
                            },
                            "wallet with type locker for user_id=2": {
                                "summary": "",
                                "value": "type=locker&available=unavailable&user_id=2"
                            }
                        }
                    },
                    {
                        "$ref": "#/components/parameters/search"
                    },
                    {
                        "parameter": "sort",
                        "name": "sort",
                        "in": "query",
                        "description": "Sort the results by field. Use '-' for descending order. Example:-id",
                        "required": false
                    },
                    {
                        "$ref": "#/components/parameters/page_limit"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "default": "No message"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "value": {
                                                        "type": "string"
                                                    },
                                                    "label": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
...
        "parameters": {
...
            "search": {
                "name": "filter[search]",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "string",
                    "default": ""
                }
            },
...
            "page_limit": {
                "name": "page_limit",
                "in": "query",
                "description": "number of items per page",
                "required": false,
                "schema": {
                    "type": "integer",
                    "default": 15
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token",
                "name": "Authorization",
                "in": "header"
            },
            "clubId": {
                "type": "apiKey",
                "description": "Club ID header",
                "name": "X-Club-Id",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": [],
            "clubId": []
        }
    ]
}